+++ source/renderer/app/api/api.js
@@ -13,6 +13,7 @@ import {
   WalletTransaction,
   TransactionTypes,
   TransactionStates,
+  TransactionWithdrawal,
 } from '../domains/WalletTransaction';
 import WalletAddress from '../domains/WalletAddress';
 
@@ -68,7 +69,7 @@ import { transferFunds } from './wallets/requests/transferFunds';
 
 // Staking
 import StakePool from '../domains/StakePool';
-import { EPOCH_LENGTH_ITN } from '../config/epochsConfig';
+import { getEpochLength } from '../config/epochsConfig';
 
 // News requests
 import { getNews } from './news/requests/getNews';
@@ -197,7 +198,7 @@ import { deleteTransaction } from './transactions/requests/deleteTransaction';
 import { WALLET_BYRON_KINDS } from '../config/walletRestoreConfig';
 import ApiError from '../domains/ApiError';
 
-const { isIncentivizedTestnet, isShelleyTestnet } = global;
+const { isIncentivizedTestnet } = global;
 
 export default class AdaApi {
   config: RequestConfig;
@@ -218,14 +219,13 @@ export default class AdaApi {
     const { isShelleyActivated } = request;
     try {
       const wallets: AdaWallets =
-        (isIncentivizedTestnet && !isShelleyTestnet) || isShelleyActivated
+        isIncentivizedTestnet || isShelleyActivated
           ? await getWallets(this.config)
           : [];
       const legacyWallets: LegacyAdaWallets = await getLegacyWallets(
         this.config
       );
       logger.debug('AdaApi::getWallets success', { wallets, legacyWallets });
-
       map(legacyWallets, legacyAdaWallet => {
         const extraLegacyWalletProps = {
           address_pool_gap: 0, // Not needed for legacy wallets
@@ -551,7 +551,7 @@ export default class AdaApi {
         passphrase: spendingPassword,
       };
 
-      if ((isIncentivizedTestnet && !isShelleyTestnet) || isShelleyActivated) {
+      if (isIncentivizedTestnet || isShelleyActivated) {
         wallet = await createWallet(this.config, {
           walletInitData,
         });
@@ -633,6 +633,7 @@ export default class AdaApi {
           },
         ],
         passphrase,
+        withdrawal: TransactionWithdrawal,
       };
 
       let response: Transaction;
@@ -694,6 +695,7 @@ export default class AdaApi {
             },
           },
         ],
+        withdrawal: TransactionWithdrawal,
       };
 
       let response: TransactionFee;
@@ -816,7 +818,7 @@ export default class AdaApi {
       const response: Promise<Array<string>> = new Promise(resolve =>
         resolve(
           generateAccountMnemonics(
-            (isIncentivizedTestnet && !isShelleyTestnet) || isShelleyActivated
+            isIncentivizedTestnet || isShelleyActivated
               ? WALLET_RECOVERY_PHRASE_WORD_COUNT
               : LEGACY_WALLET_RECOVERY_PHRASE_WORD_COUNT
           )
@@ -1583,7 +1585,7 @@ export default class AdaApi {
           : get(sync_progress, 'progress.quantity', 0);
       const nextEpochNumber = get(nextEpoch, 'epoch_number', null);
       const nextEpochStartTime = get(nextEpoch, 'epoch_start_time', '');
-
+      const epochLength = getEpochLength();
       // extract relevant data before sending to NetworkStatusStore
       return {
         syncProgress,
@@ -1610,7 +1612,7 @@ export default class AdaApi {
               epochNumber: nextEpochNumber ? nextEpochNumber + 1 : null,
               epochStart: nextEpochStartTime
                 ? moment(nextEpochStartTime)
-                    .add(EPOCH_LENGTH_ITN, 'seconds')
+                    .add(epochLength, 'seconds')
                     .toISOString()
                 : '',
             }
@@ -1672,6 +1674,7 @@ export default class AdaApi {
         decentralization_level: decentralizationLevel,
         desired_pool_number: desiredPoolNumber,
         minimum_utxo_value: minimumUtxoValue,
+        hardfork_at: hardforkAt,
       } = networkParameters;
       const blockchainStartTime = moment(blockchain_start_time).valueOf();
 
@@ -1685,6 +1688,7 @@ export default class AdaApi {
         decentralizationLevel,
         desiredPoolNumber,
         minimumUtxoValue,
+        hardforkAt: hardforkAt || null,
       };
     } catch (error) {
       logger.error('AdaApi::getNetworkParameters error', { error });