+++ source/renderer/app/components/loading/syncing-connecting/SyncingConnectingStatus.js
@@ -65,16 +65,25 @@ const messages = defineMessages({
     defaultMessage: '!!!TLS certificate is not valid, please restart Daedalus.',
     description: 'The TLS cert is not valid and Daedalus should be restarted',
   },
+  verifyingBlockchain: {
+    id: 'loading.screen.verifyingBlockchainMessage',
+    defaultMessage:
+      '!!!Verifying the blockchain ({verificationProgress}% complete)',
+    description:
+      'Message "Verifying the blockchain (65% complete) ..." on the loading screen.',
+  },
 });
 
 type Props = {
   cardanoNodeState: ?CardanoNodeState,
+  verificationProgress: number,
   hasLoadedCurrentLocale: boolean,
   hasBeenConnected: boolean,
   isTlsCertInvalid: boolean,
   isConnected: boolean,
   isNodeStopping: boolean,
   isNodeStopped: boolean,
+  isVerifyingBlockchain: boolean,
 };
 
 export default class SyncingConnectingStatus extends Component<Props> {
@@ -139,6 +148,8 @@ export default class SyncingConnectingStatus extends Component<Props> {
       isNodeStopped,
       isTlsCertInvalid,
       hasLoadedCurrentLocale,
+      isVerifyingBlockchain,
+      verificationProgress,
     } = this.props;
 
     if (!hasLoadedCurrentLocale) return null;
@@ -159,7 +170,11 @@ export default class SyncingConnectingStatus extends Component<Props> {
     return (
       <div className={componentStyles}>
         <h1 className={headlineStyles}>
-          {intl.formatMessage(this._getConnectingMessage())}
+          {isVerifyingBlockchain
+            ? intl.formatMessage(messages.verifyingBlockchain, {
+                verificationProgress,
+              })
+            : intl.formatMessage(this._getConnectingMessage())}
         </h1>
       </div>
     );