+++ source/renderer/app/components/wallet/wallet-restore/WalletTypeDialog.js
@@ -164,8 +164,6 @@ type State = {
   [key: HardwareWalletAcceptance]: boolean,
 };
 
-const { isIncentivizedTestnet, isShelleyTestnet } = global;
-
 export default class WalletTypeDialog extends Component<Props, State> {
   static contextTypes = {
     intl: intlShape.isRequired,
@@ -198,7 +196,9 @@ export default class WalletTypeDialog extends Component<Props, State> {
         }
         return {
           key: kind,
-          disabled: !isShelleyActivated && kind.includes('Shelley'),
+          disabled:
+            (!isShelleyActivated && kind.includes('Shelley')) ||
+            kind === WALLET_YOROI_KINDS.SHELLEY_15_WORD,
           label: <FormattedHTMLMessage {...msg} />,
           selected: value === kind,
           onChange: () => this.props.onSetWalletKind(kind, kindParam),
@@ -223,12 +223,7 @@ export default class WalletTypeDialog extends Component<Props, State> {
     if (!walletKind) return true;
     if (walletKind === WALLET_KINDS.DAEDALUS && !walletKindDaedalus)
       return true;
-    if (
-      walletKind === WALLET_KINDS.YOROI &&
-      !walletKindYoroi &&
-      isIncentivizedTestnet
-    )
-      return true;
+    if (walletKind === WALLET_KINDS.YOROI && !walletKindYoroi) return true;
     return (
       walletKind === WALLET_KINDS.HARDWARE &&
       (!walletKindHardware ||
@@ -284,8 +279,7 @@ export default class WalletTypeDialog extends Component<Props, State> {
               walletKindDaedalus,
               WALLET_KINDS.DAEDALUS
             )}
-          {isIncentivizedTestnet &&
-            walletKind === WALLET_KINDS.YOROI &&
+          {walletKind === WALLET_KINDS.YOROI &&
             this.getWalletKind(
               isShelleyActivated,
               WALLET_YOROI_KINDS,
@@ -293,7 +287,7 @@ export default class WalletTypeDialog extends Component<Props, State> {
               walletKindYoroi,
               WALLET_KINDS.YOROI
             )}
-          {isShelleyTestnet && walletKind === WALLET_KINDS.HARDWARE && (
+          {walletKind === WALLET_KINDS.HARDWARE && (
             <Fragment>
               {this.getWalletKind(
                 isShelleyActivated,