+++ source/renderer/app/containers/MainLayout.js
@@ -35,10 +35,16 @@ export default class MainLayout extends Component<InjectedContainerProps> {
 
   render() {
     const { actions, stores } = this.props;
-    const { sidebar, profile, app, wallets: walletsStore, staking } = stores;
+    const {
+      sidebar,
+      profile,
+      app,
+      wallets: walletsStore,
+      networkStatus,
+    } = stores;
     const activeWallet = walletsStore.active;
     const activeWalletId = activeWallet ? activeWallet.id : null;
-    const { isShelleyActivated } = staking;
+    const { isShelleyActivated } = networkStatus;
     const { currentTheme } = profile;
     const {
       environment: { network, isDev },
@@ -80,7 +86,6 @@ export default class MainLayout extends Component<InjectedContainerProps> {
         menus={sidebarMenus}
         isShowingSubMenus={sidebar.isShowingSubMenus}
         isIncentivizedTestnet={global.isIncentivizedTestnet}
-        isShelleyTestnet={global.isShelleyTestnet}
         isShelleyActivated={isShelleyActivated}
         categories={sidebar.CATEGORIES}
         activeSidebarCategory={sidebar.activeSidebarCategory}